Initial script upload
ccl_source_funs contains a (partial) set of generalized functions to streamline translation of country checklists to Map of Life. Functions assume that each row in the data table contains a species name and a geographic ID indicating that the species is known to occur within that geographic area. May also include a column with alternative taxonomic name, a column detailing the source of the record, or other information accompanying the checklist.

Country_checklist_mobilization is a workflow that calls functions from ccl_source_funs to read checklist data into R, harmonize the taxonomy, calculate country richness, and match raw records and richness to MOL geo IDs for GADM level 0 administrative units.